changed parameters in the fetch data function to solve the error of the information not updating (it is a test), also cleaned up code and improved readability
